Change `body` to `description` in release pull request creation
This commit modifies the `cd.yml` workflow. While creating a new pull request, we were referencing the `body` of the GitHub release but it appears to be inaccurate. We corrected it to refer to `description`, which tends to contain the data necessary for seamless pull request creation during the deployment process. This eliminates the potential for unnecessary failure during the PR creation.